unify sdk and cli upload
